refactor: remove useless functions in utils
