+++ src/application/session/aggregates/session.ts
@@ -2,7 +2,7 @@ import { Pair, Entity, Repository } from '../entity'
 import { EventBus, EventBusImpl } from '../events/eventBus'
 import { EventType, UserJoinedEvent, UserLeftSessionEvent } from '../events/events'
 import { Chat, ChatImpl } from './chat'
-import { User, UserRepository } from './user'
+import { User, UserRepository } from '../user'
 import { Video, VideoImpl } from './video'
 
 export class SessionId {
