+++ src/presentation/presentation/deserialization/userDeserializer.ts
@@ -1,4 +1,4 @@
-import { UserId, User } from '../../../application/session/aggregates/user'
+import { UserId, User } from '../../../application/session/user'
 import { AbstractDeserializer } from './deserializer'
 
 class UserIdDeserializer extends AbstractDeserializer<UserId> {
