+++ src/presentation/registerCommands.ts
